requirements.txt
scikit-learn
pandas
python-telegram-bot --pre